finally slayed the publicprofile
