--- client/src/components/PublicProfile/PublicProfile-container.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import PublicProfileView from './PublicProfile-view';
-
-export default function PublicProfileContainer(props) {
-  return <PublicProfileView {...props} />;
-}
