+++ client/src/components/PublicProfile/PublicProfile-container.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import PublicProfileView, { PublicProfileViewProps } from './PublicProfile-view';
+
+type PublicProfileContainerProps = PublicProfileViewProps;
+
+export default function PublicProfileContainer(props: PublicProfileContainerProps): JSX.Element {
+  return <PublicProfileView {...props} />;
+}
